Refactor rating system and adjust UI components
This commit introduces several changes:

- **Rating System:**
    - `RatingBottomSheet` and `StarRatingBar`:
        - Changed `onRatingChange` parameter from `(Float) -> Unit` to `(Int) -> Unit`.
        - Star selection logic now uses `ceil(newRating).toInt()` to round up to the nearest whole star.
        - Star filling logic changed to fully fill a star if `rating >= i`.
    - `MyRequestScreen`: Adapted `onRatingChange` to cast the `Int` to `Float` when calling the interaction listener.
- **RequestCard:**
    - Added `enableSecondaryBackgroundColor` parameter to the secondary `AppButton`, setting it to `Theme.colors.shade.quinary`.
- **ServiceTypeCard:**
    - Increased the `Spacer` width between the icon and text from `8.dp` to `12.dp`.

+++ app/src/main/java/com/paris_2/san3a/presentation/screen/requests/customer/MyRequestScreen.kt
@@ -51,7 +51,7 @@ private fun MyRequestScreenContent(
         rating = state.myRequestCustomerUiState.rating,
         onDismiss = myRequestCustomerInteractionListener::onRatingDismiss,
         onRatingChange = { rating ->
-            myRequestCustomerInteractionListener.onRatingChange(rating)
+            myRequestCustomerInteractionListener.onRatingChange(rating.toFloat())
         },
         onAddRating = {
             myRequestCustomerInteractionListener.onRatingCraftsMan()
